update complete page, create user history hook
